+++ CHANGELOG.md
@@ -4,8 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2020-11-14
+### Added
+- IS_NULL, NOT_NULL and NOT_LIKE comparators to enum.
+
+### Removed
+- NOT comparator.
+
+### Refactored
+- PSR12 Coding standard.
+
 ## 1.0.0 - 2020-04-09
 ### Added
 - Initial implementation of the package.
 
-[1.0.0 > Unreleased]: https://github.com/grizz-it/search/compare/1.0.0...HEAD
+- [2.0.0 > Unreleased](https://github.com/grizz-it/search/compare/2.0.0...HEAD)
+- [1.0.0 > 2.0.0](https://github.com/grizz-it/search/compare/1.0.0...2.0.0)